+++ lib/processor/dedupe.go
@@ -32,7 +32,6 @@ import (
 	"github.com/Jeffail/benthos/lib/response"
 	"github.com/Jeffail/benthos/lib/types"
 	"github.com/Jeffail/benthos/lib/util/text"
-	"github.com/Jeffail/gabs"
 )
 
 //------------------------------------------------------------------------------
@@ -61,8 +60,6 @@ dedupe:
   key: ${!metadata:kafka_key}-${!json_field:id}
 ` + "```" + `
 
-The ` + "`json_paths`" + ` field is deprecated.
-
 Caches should be configured as a resource, for more information check out the
 [documentation here](../caches).`,
 	}
@@ -72,12 +69,11 @@ Caches should be configured as a resource, for more information check out the
 
 // DedupeConfig contains configuration fields for the Dedupe processor.
 type DedupeConfig struct {
-	Cache          string   `json:"cache" yaml:"cache"`
-	HashType       string   `json:"hash" yaml:"hash"`
-	Parts          []int    `json:"parts" yaml:"parts"` // message parts to hash
-	Key            string   `json:"key" yaml:"key"`
-	JSONPaths      []string `json:"json_paths" yaml:"json_paths"`
-	DropOnCacheErr bool     `json:"drop_on_err" yaml:"drop_on_err"`
+	Cache          string `json:"cache" yaml:"cache"`
+	HashType       string `json:"hash" yaml:"hash"`
+	Parts          []int  `json:"parts" yaml:"parts"` // message parts to hash
+	Key            string `json:"key" yaml:"key"`
+	DropOnCacheErr bool   `json:"drop_on_err" yaml:"drop_on_err"`
 }
 
 // NewDedupeConfig returns a DedupeConfig with default values.
@@ -87,7 +83,6 @@ func NewDedupeConfig() DedupeConfig {
 		HashType:       "none",
 		Parts:          []int{0}, // only consider the 1st part
 		Key:            "",
-		JSONPaths:      []string{},
 		DropOnCacheErr: true,
 	}
 }
@@ -147,7 +142,6 @@ type Dedupe struct {
 
 	cache      types.Cache
 	hasherFunc hasherFunc
-	jPaths     []string
 
 	mCount     metrics.StatCounter
 	mErrJSON   metrics.StatCounter
@@ -162,14 +156,6 @@ type Dedupe struct {
 func NewDedupe(
 	conf Config, mgr types.Manager, log log.Modular, stats metrics.Type,
 ) (Type, error) {
-	if len(conf.Dedupe.JSONPaths) > 0 {
-		log.Warnln(
-			"WARNING: The 'json_paths' field of the 'dedupe' processor is" +
-				" deprecated, instead use the 'key' field with function" +
-				" interpolation, e.g. '${!json_field:foo.bar}'",
-		)
-	}
-
 	c, err := mgr.GetCache(conf.Dedupe.Cache)
 	if err != nil {
 		return nil, err
@@ -193,7 +179,6 @@ func NewDedupe(
 
 		cache:      c,
 		hasherFunc: hFunc,
-		jPaths:     conf.Dedupe.JSONPaths,
 
 		mCount:     stats.GetCounter("processor.dedupe.count"),
 		mErrJSON:   stats.GetCounter("processor.dedupe.error.json_parse"),
@@ -222,50 +207,8 @@ func (d *Dedupe) ProcessMessage(msg types.Message) ([]types.Message, types.Respo
 	if len(key) > 0 {
 		hasher.Write(key)
 		extractedHash = true
-	}
-
-	for _, index := range d.conf.Dedupe.Parts {
-		if len(d.jPaths) > 0 {
-			// Attempt to add JSON fields from part to hash.
-			jPart, err := msg.GetJSON(index)
-			if err != nil {
-				d.mErrJSON.Incr(1)
-				d.mDropped.Incr(1)
-				d.log.Errorf("JSON Parse error: %v\n", err)
-				continue
-			}
-
-			var gPart *gabs.Container
-			if gPart, err = gabs.Consume(jPart); err != nil {
-				d.mErrJSON.Incr(1)
-				d.mDropped.Incr(1)
-				d.log.Errorf("JSON Parse error: %v\n", err)
-				continue
-			}
-
-			for _, jPath := range d.jPaths {
-				gTarget := gPart.Path(jPath)
-				if gTarget.Data() == nil {
-					continue
-				}
-
-				var hashBytes []byte
-				switch t := gTarget.Data().(type) {
-				case string:
-					hashBytes = []byte(t)
-				default:
-					hashBytes = gTarget.Bytes()
-				}
-
-				if _, err := hasher.Write(hashBytes); nil != err {
-					d.mErrHash.Incr(1)
-					d.mDropped.Incr(1)
-					d.log.Errorf("Hash error: %v\n", err)
-				} else {
-					extractedHash = true
-				}
-			}
-		} else if len(key) == 0 {
+	} else {
+		for _, index := range d.conf.Dedupe.Parts {
 			// Attempt to add whole part to hash.
 			if partBytes := msg.Get(index); partBytes != nil {
 				if _, err := hasher.Write(msg.Get(index)); nil != err {